requirements.txt
ultralytics>=8.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
opencv-python>=4.6.0
Pillow>=9.0.0
torch>=2.0.0
requests>=2.28.0
